requirements.txt
flask
flask-sqlalchemy
flask-login
pyodbc
python-dotenv
